+++ tdms/CMakeLists.txt
@@ -67,6 +67,8 @@ set(SOURCES
     src/fields/split.cpp
     src/fields/td_field_exporter_2d.cpp
     src/argument_parser.cpp
+    src/array_init.cpp
+    src/arrays.cpp
     src/dimensions.cpp
     src/fdtd_grid_initialiser.cpp
     src/grid_labels.cpp
@@ -81,10 +83,8 @@ set(SOURCES
     src/shapes.cpp
     src/simulation_parameters.cpp
     src/source.cpp
-    src/tensor_init.cpp
     src/timer.cpp
     src/utils.cpp
-    src/vector_collection.cpp
     matlabio/matlabio.cpp
     )
