+++ tdms/include/iterator.h
@@ -1,6 +1,6 @@
+#include "arrays.h"
 #include "complex"
 #include "field.h"
-#include "vector_collection.h"
 #include "grid_labels.h"
