--- tdms/include/tensor_collection.h
@@ -1,18 +1,0 @@
-#include "utils.h"
-
-
-class XYZTensor3D {
-public:
-  double ***x = nullptr;
-  double ***y = nullptr;
-  double ***z = nullptr;
-
-  double*** operator[] (char c) const{
-    switch (c) {
-      case 'x': return x;
-      case 'y': return y;
-      case 'z': return z;
-      default: throw std::runtime_error("Have no element " + to_string(c));
-    }
-  }
-};
