+++ tdms/src/arrays.cpp
@@ -1,7 +1,7 @@
-#include <utility>
 #include <iostream>
+#include <utility>
+#include "arrays.h"
 #include "globals.h"
-#include "vector_collection.h"
 #include "utils.h"
